Readme update and resource management optimization
Added links to the version and download count of NuGet packages in Readme.md. Introduced support for CancellationToken in the GetCountAsync() method in IDatabasesRepository.cs. Removed unused `using` directives and changed database context creation to use `using` statements in DatabasesRepository.cs. Updated the implementation of the Dispose() method, adding the Dispose(bool disposing) method and calling GC.SuppressFinalize(this).